Share moved to own atom component
