+++ packages/ui-lib/src/Layouts/molecules/UtilityHeader.tsx
@@ -1,10 +1,10 @@
-import React, { useState, useCallback, useEffect } from "react";
+import React from "react";
 import styled, { css } from "styled-components";
 import {Link} from 'react-router-dom';
 
 import Icon from "../../Icons/Icon";
-import { IUtilityHeader, IUtilityHeaderLinkBack, IUtilityHeaderLinkShare } from "..";
-import { useCopyToClipboard } from "../../hooks";
+import { IUtilityHeader, IUtilityHeaderLinkBack } from "..";
+import UtilityHeaderShare from "../atoms/UtilityHeaderShare";
 
 const Container = styled.div`
   max-width: var(--max-content-width);
@@ -80,49 +80,6 @@ const BackLink = styled(Link)<{iconInGutter: boolean, showDivider: boolean}>`
   `}
 `;
 
-const ExtraActionIcon = styled.div`
-  display: flex;
-  width: 16px;
-  height: 16px;
-  justify-content: center;
-  align-items: center;
-`;
-const ExtraAction = styled.button`
-  position: relative;
-  display: flex;
-  padding: 0;
-  align-items: center;
-  gap: 8px;
-  color: var(--grey-10);
-  text-align: center;
-  font-family: var(--font-ui);
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  border: none;
-  background: none;
-  text-decoration: none;
-  transition: color 0.25s ease;
-  cursor: pointer;
-
-  ${ExtraActionIcon}{
-    svg * {
-      transition: stroke 0.25s ease;
-    }
-  }
-  
-  &:hover {
-    color: var(--grey-12);
-    ${ExtraActionIcon}{
-      svg * {
-        stroke: var(--grey-12);
-      }
-    }
-  }
-
-`;
-
 const Breadcrumbs = styled.div`
   display: inline-flex;
   align-items: center;
@@ -185,33 +142,7 @@ const RightArea = styled.div`
   justify-content: right;
 `;
 
-const UtilityHeaderShare : React.FC<IUtilityHeaderLinkShare> = ({show, link, label = 'Share', copiedLabel = 'Copied'}) => {
-  
-  const [ copyActionText, setCopyActionText ] = useState<string>(label);
-  const {copyToClipboard} = useCopyToClipboard();
-  
-  const clickHandlerShareLink = useCallback(() => {
-    copyToClipboard( link ? link : window.location.href);
-    setCopyActionText(copiedLabel);
-    setTimeout(() => setCopyActionText(copyActionText), 2000);
-  }, [link, copiedLabel, copyActionText, copyToClipboard]);
 
-  useEffect(() => {
-    setCopyActionText(label);
-  }, [label]);
-  
-  if(!show){ return null; }
-  
-  return(
-    <ExtraAction onClick={ clickHandlerShareLink }>
-      <ExtraActionIcon>
-        <Icon icon="Link" size={16} color="grey-10" />
-      </ExtraActionIcon>
-      {copyActionText}
-    </ExtraAction>
-  );
-  
-};
 
 interface IUtilityHeaderLinkBackInstance extends IUtilityHeaderLinkBack {
   iconInGutter: boolean;
